Revert "Add currentwar support"
