+++ CocNET/CocNET.Test/packages.config
@@ -1,11 +1,9 @@
 ﻿<?xml version="1.0" encoding="utf-8"?>
 <packages>
-  <package id="Funq" version="1.0.0" targetFramework="net452" />
   <package id="Microsoft.Bcl" version="1.1.10" targetFramework="net452" />
   <package id="Microsoft.Bcl.Build" version="1.0.14" targetFramework="net452" />
   <package id="Microsoft.Net.Http" version="2.2.29" targetFramework="net452" />
   <package id="morelinq" version="1.4.0" targetFramework="net452" />
   <package id="Newtonsoft.Json" version="8.0.3" targetFramework="net452" />
-  <package id="NUnit" version="3.6.1" targetFramework="net452" />
   <package id="RestSharp" version="105.2.3" targetFramework="net452" />
 </packages>
