+++ CocNET/CocNET/CocNET.csproj
@@ -11,8 +11,6 @@
     <AssemblyName>CocNET</AssemblyName>
     <TargetFrameworkVersion>v4.5.2</TargetFrameworkVersion>
     <FileAlignment>512</FileAlignment>
-    <SolutionDir Condition="$(SolutionDir) == '' Or $(SolutionDir) == '*Undefined*'">..\</SolutionDir>
-    <RestorePackages>true</RestorePackages>
   </PropertyGroup>
   <PropertyGroup Condition=" '$(Configuration)|$(Platform)' == 'Debug|AnyCPU' ">
     <DebugSymbols>true</DebugSymbols>
@@ -32,10 +30,6 @@
     <WarningLevel>4</WarningLevel>
   </PropertyGroup>
   <ItemGroup>
-    <Reference Include="Funq, Version=1.0.0.0, Culture=neutral, PublicKeyToken=95c13cd3fe8e976a, processorArchitecture=MSIL">
-      <HintPath>..\packages\Funq.1.0.0\lib\net40\Funq.dll</HintPath>
-      <Private>True</Private>
-    </Reference>
     <Reference Include="MoreLinq, Version=1.4.18916.0, Culture=neutral, PublicKeyToken=384d532d7e88985d, processorArchitecture=MSIL">
       <HintPath>..\packages\morelinq.1.4.0\lib\net35\MoreLinq.dll</HintPath>
       <Private>True</Private>
@@ -70,21 +64,13 @@
   <ItemGroup>
     <Compile Include="CocCore.cs" />
     <Compile Include="Includes\Locations.cs" />
-    <Compile Include="Interfaces\ICocCoreClans.cs" />
-    <Compile Include="Interfaces\ICocCoreLeagues.cs" />
-    <Compile Include="Interfaces\ICocCoreLocations.cs" />
-    <Compile Include="Interfaces\ICocCorePlayers.cs" />
+    <Compile Include="Interfaces\ICocCore.cs" />
     <Compile Include="Methods\Request.cs" />
     <Compile Include="Methods\UrlBuilder.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
-    <Compile Include="Services\CocCoreClans.cs" />
-    <Compile Include="Services\CocCoreLeagues.cs" />
-    <Compile Include="Services\CocCoreLocations.cs" />
-    <Compile Include="Services\CocCorePlayers.cs" />
     <Compile Include="Types\Clans\Opponent.cs" />
     <Compile Include="Types\Clans\WarLog.cs" />
     <Compile Include="Types\Clans\WarLogs.cs" />
-    <Compile Include="Types\Other\Game\Achievement.cs" />
     <Compile Include="Types\Other\BadRequest.cs" />
     <Compile Include="Types\Clans\Clan.cs" />
     <Compile Include="Types\Clans\ClanRanking.cs" />
@@ -92,14 +78,9 @@
     <Compile Include="Types\Leagues\League.cs" />
     <Compile Include="Types\Locations\Location.cs" />
     <Compile Include="Types\Locations\Locations.cs" />
-    <Compile Include="Types\Other\Game\Hero.cs" />
-    <Compile Include="Types\Other\Game\Spell.cs" />
-    <Compile Include="Types\Other\Game\Troop.cs" />
     <Compile Include="Types\Other\Member.cs" />
     <Compile Include="Types\Other\Members.cs" />
     <Compile Include="Types\Other\Paging.cs" />
-    <Compile Include="Types\Other\Game\Season.cs" />
-    <Compile Include="Types\Other\Statistics\LegendStatistics.cs" />
     <Compile Include="Types\Players\Player.cs" />
     <Compile Include="Types\Players\PlayerRanking.cs" />
     <Compile Include="Types\Other\Ranking.cs" />
@@ -118,13 +99,6 @@
   <Target Name="EnsureBclBuildImported" BeforeTargets="BeforeBuild" Condition="'$(BclBuildImported)' == ''">
     <Error Condition="!Exists('..\packages\Microsoft.Bcl.Build.1.0.14\tools\Microsoft.Bcl.Build.targets')" Text="This project references NuGet package(s) that are missing on this computer. Enable NuGet Package Restore to download them.  For more information, see http://go.microsoft.com/fwlink/?LinkID=317567." HelpKeyword="BCLBUILD2001" />
     <Error Condition="Exists('..\packages\Microsoft.Bcl.Build.1.0.14\tools\Microsoft.Bcl.Build.targets')" Text="The build restored NuGet packages. Build the project again to include these packages in the build. For more information, see http://go.microsoft.com/fwlink/?LinkID=317568." HelpKeyword="BCLBUILD2002" />
-  </Target>
-  <Import Project="$(SolutionDir)\.nuget\NuGet.targets" Condition="Exists('$(SolutionDir)\.nuget\NuGet.targets')" />
-  <Target Name="EnsureNuGetPackageBuildImports" BeforeTargets="PrepareForBuild">
-    <PropertyGroup>
-      <ErrorText>This project references NuGet package(s) that are missing on this computer. Enable NuGet Package Restore to download them.  For more information, see http://go.microsoft.com/fwlink/?LinkID=322105. The missing file is {0}.</ErrorText>
-    </PropertyGroup>
-    <Error Condition="!Exists('$(SolutionDir)\.nuget\NuGet.targets')" Text="$([System.String]::Format('$(ErrorText)', '$(SolutionDir)\.nuget\NuGet.targets'))" />
   </Target>
   <!-- To modify your build process, add your task inside one of the targets below and uncomment it. 
        Other similar extension points exist, see Microsoft.Common.targets.
