--- CocNET/CocNET/Types/Other/Game/Spell.cs
@@ -1,19 +1,0 @@
-﻿using Newtonsoft.Json;
-using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-
-namespace CocNET.Types.Other.Game
-{
-    public class Spell : BadRequest
-    {
-        [JsonProperty("name")]
-        public string Name { get; set; }
-        [JsonProperty("level")]
-        public int Level { get; set; }
-        [JsonProperty("maxLevel")]
-        public int MaxLevel { get; set; }
-    }
-}
